+++ src/Element/Queries/AddressQuery.php
@@ -324,7 +324,12 @@ class AddressQuery extends ElementQuery implements NestedElementQueryInterface
 
     public function shouldApplyNestedElementParams(): bool
     {
-        return ! empty($this->fieldId);
+        return isset($this->fieldId);
+    }
+
+    protected function mustHaveField(): bool
+    {
+        return false;
     }
 
     /** @param array<string, mixed> $config */
@@ -357,7 +362,7 @@ class AddressQuery extends ElementQuery implements NestedElementQueryInterface
         $this->beforeQuery(static function (self $addressQuery) {
             $addressQuery->normalizeNestedElementParams($addressQuery);
 
-            if (empty($addressQuery->fieldId) && (isset($addressQuery->primaryOwnerId) || isset($addressQuery->ownerId))) {
+            if (! isset($addressQuery->fieldId) && (isset($addressQuery->primaryOwnerId) || isset($addressQuery->ownerId))) {
                 // User addresses don't get rows in the elements_owners table
                 if (! $addressQuery->primaryOwnerId && ! $addressQuery->ownerId) {
                     throw new QueryAbortedException;