+++ src/Field/Field.php
@@ -352,12 +352,12 @@ abstract class Field extends Component implements Actionable, FieldInterface, Ic
         return Query::TYPE_TEXT;
     }
 
-    public static function modifyQuery(Builder $query, array $instances, mixed $value): Builder
+    public static function modifyQuery(Builder $query, array $instances, mixed $value): void
     {
         $valueSql = static::valueSql($instances);
 
         if ($valueSql === null) {
-            return $query;
+            return;
         }
 
         $caseInsensitive = false;
@@ -367,7 +367,7 @@ abstract class Field extends Component implements Actionable, FieldInterface, Ic
             $value = $value['value'];
         }
 
-        return $query->whereParam(
+        $query->whereParam(
             column: $valueSql,
             param: $value,
             caseInsensitive: $caseInsensitive,